+++ engine/transpiler/src/defaults.ts
@@ -13,9 +13,11 @@ export const defaultUtilitySelectorFn: UtilitySelectorFn = ({
 	value,
 	modifiers,
 }) => {
-	return `._${[...modifiers, name, value].filter(Boolean).join("\\:")}`;
+	return `._${[...modifiers, name, ...(value === "default" ? [] : [value])]
+		.filter(Boolean)
+		.join("\\:")}`;
 };
 
 export const defaultVariableNameFn: VariableNameFn = ({ name }) => {
-	return name;
+	return name.replace(/^\.+|\.+$/g, "").replace(/\.+/g, "--");
 };