+++ theme/src/variables/useScalePowers.ts
@@ -6,7 +6,10 @@ import type {
 	Variable,
 } from "@styleframe/core";
 import { isRef } from "@styleframe/core";
-import { defaultScalePowerValues } from "../constants";
+
+export const defaultScalePowerValues: readonly number[] = [
+	-2, -1, 0, 1, 2, 3, 4, 5,
+] as const;
 
 export function useScalePowers<T extends readonly number[]>(
 	s: Styleframe,